requirements.txt
langchain==0.0.140
nltk==3.8.1
openai==0.27.2
pinecone_client==2.2.1
python-dotenv==1.0.0
PyYAML==6.0
fastapi==0.68.2
uvicorn==0.15.0
gptrim==0.1.1
googlemaps==4.10.0
jinja2==3.1.2